Fix: update tracker contract from 5-tuple to 6-tuple (upper_body)
- tracker.py: remove int() casts from upper body coordinates for type consistency with ankle pixels (float tuples)
- main.py: update process_frame unpacking from 5-tuple to 6-tuple, adding upper_body
- tests/test_e2e.py: update TestTrackerSmoke to validate 6-tuple contract; update _build_mock_side_effect and _run_pipeline_loop to return/consume 6-tuples

+++ tracker.py
@@ -114,12 +114,12 @@ class PlayerTracker:
 
                 upper_body = {
                     "shoulders_px": (
-                        (int(x1 + l_shoulder.x * roi_width), int(y1 + l_shoulder.y * roi_height)),
-                        (int(x1 + r_shoulder.x * roi_width), int(y1 + r_shoulder.y * roi_height)),
+                        (x1 + l_shoulder.x * roi_width, y1 + l_shoulder.y * roi_height),
+                        (x1 + r_shoulder.x * roi_width, y1 + r_shoulder.y * roi_height),
                     ),
                     "wrists_px": (
-                        (int(x1 + l_wrist.x * roi_width), int(y1 + l_wrist.y * roi_height)),
-                        (int(x1 + r_wrist.x * roi_width), int(y1 + r_wrist.y * roi_height)),
+                        (x1 + l_wrist.x * roi_width, y1 + l_wrist.y * roi_height),
+                        (x1 + r_wrist.x * roi_width, y1 + r_wrist.y * roi_height),
                     ),
                 }
